+++ include/tensorflowfake_module.h
@@ -4,35 +4,35 @@
 #include "runtime/importobj.h"
 #include "tensorflowfake_func.h"
 
-namespace SSVM {
+namespace WasmEdge {
 namespace Host {
 
-class SSVMTensorflowFakeModule : public Runtime::ImportObject {
+class WasmEdgeTensorflowFakeModule : public Runtime::ImportObject {
 public:
-  SSVMTensorflowFakeModule() : ImportObject("ssvm_tensorflow") {
-    addHostFunc("ssvm_tensorflow_create_session",
-                std::make_unique<SSVMTensorflowFakeCreateSession>());
-    addHostFunc("ssvm_tensorflow_delete_session",
-                std::make_unique<SSVMTensorflowFakeDeleteSession>());
-    addHostFunc("ssvm_tensorflow_run_session",
-                std::make_unique<SSVMTensorflowFakeRunSession>());
-    addHostFunc("ssvm_tensorflow_get_output_tensor",
-                std::make_unique<SSVMTensorflowFakeGetOutputTensor>());
-    addHostFunc("ssvm_tensorflow_get_tensor_len",
-                std::make_unique<SSVMTensorflowFakeGetTensorLen>());
-    addHostFunc("ssvm_tensorflow_get_tensor_data",
-                std::make_unique<SSVMTensorflowFakeGetTensorData>());
-    addHostFunc("ssvm_tensorflow_append_input",
-                std::make_unique<SSVMTensorflowFakeAppendInput>());
-    addHostFunc("ssvm_tensorflow_append_output",
-                std::make_unique<SSVMTensorflowFakeAppendOutput>());
-    addHostFunc("ssvm_tensorflow_clear_input",
-                std::make_unique<SSVMTensorflowFakeClearInput>());
-    addHostFunc("ssvm_tensorflow_clear_output",
-                std::make_unique<SSVMTensorflowFakeClearOutput>());
+  WasmEdgeTensorflowFakeModule() : ImportObject("wasmedge_tensorflow") {
+    addHostFunc("wasmedge_tensorflow_create_session",
+                std::make_unique<WasmEdgeTensorflowFakeCreateSession>());
+    addHostFunc("wasmedge_tensorflow_delete_session",
+                std::make_unique<WasmEdgeTensorflowFakeDeleteSession>());
+    addHostFunc("wasmedge_tensorflow_run_session",
+                std::make_unique<WasmEdgeTensorflowFakeRunSession>());
+    addHostFunc("wasmedge_tensorflow_get_output_tensor",
+                std::make_unique<WasmEdgeTensorflowFakeGetOutputTensor>());
+    addHostFunc("wasmedge_tensorflow_get_tensor_len",
+                std::make_unique<WasmEdgeTensorflowFakeGetTensorLen>());
+    addHostFunc("wasmedge_tensorflow_get_tensor_data",
+                std::make_unique<WasmEdgeTensorflowFakeGetTensorData>());
+    addHostFunc("wasmedge_tensorflow_append_input",
+                std::make_unique<WasmEdgeTensorflowFakeAppendInput>());
+    addHostFunc("wasmedge_tensorflow_append_output",
+                std::make_unique<WasmEdgeTensorflowFakeAppendOutput>());
+    addHostFunc("wasmedge_tensorflow_clear_input",
+                std::make_unique<WasmEdgeTensorflowFakeClearInput>());
+    addHostFunc("wasmedge_tensorflow_clear_output",
+                std::make_unique<WasmEdgeTensorflowFakeClearOutput>());
   }
-  ~SSVMTensorflowFakeModule() = default;
+  ~WasmEdgeTensorflowFakeModule() = default;
 };
 
 } // namespace Host
-} // namespace SSVM
+} // namespace WasmEdge