+++ lib/tensorflow_module.cpp
@@ -4,31 +4,32 @@
 
 #include <memory>
 
-namespace SSVM {
+namespace WasmEdge {
 namespace Host {
 
-SSVMTensorflowModule::SSVMTensorflowModule() : ImportObject("ssvm_tensorflow") {
-  addHostFunc("ssvm_tensorflow_create_session",
-              std::make_unique<SSVMTensorflowCreateSession>());
-  addHostFunc("ssvm_tensorflow_delete_session",
-              std::make_unique<SSVMTensorflowDeleteSession>());
-  addHostFunc("ssvm_tensorflow_run_session",
-              std::make_unique<SSVMTensorflowRunSession>());
-  addHostFunc("ssvm_tensorflow_get_output_tensor",
-              std::make_unique<SSVMTensorflowGetOutputTensor>());
-  addHostFunc("ssvm_tensorflow_get_tensor_len",
-              std::make_unique<SSVMTensorflowGetTensorLen>());
-  addHostFunc("ssvm_tensorflow_get_tensor_data",
-              std::make_unique<SSVMTensorflowGetTensorData>());
-  addHostFunc("ssvm_tensorflow_append_input",
-              std::make_unique<SSVMTensorflowAppendInput>());
-  addHostFunc("ssvm_tensorflow_append_output",
-              std::make_unique<SSVMTensorflowAppendOutput>());
-  addHostFunc("ssvm_tensorflow_clear_input",
-              std::make_unique<SSVMTensorflowClearInput>());
-  addHostFunc("ssvm_tensorflow_clear_output",
-              std::make_unique<SSVMTensorflowClearOutput>());
+WasmEdgeTensorflowModule::WasmEdgeTensorflowModule()
+    : ImportObject("wasmedge_tensorflow") {
+  addHostFunc("wasmedge_tensorflow_create_session",
+              std::make_unique<WasmEdgeTensorflowCreateSession>());
+  addHostFunc("wasmedge_tensorflow_delete_session",
+              std::make_unique<WasmEdgeTensorflowDeleteSession>());
+  addHostFunc("wasmedge_tensorflow_run_session",
+              std::make_unique<WasmEdgeTensorflowRunSession>());
+  addHostFunc("wasmedge_tensorflow_get_output_tensor",
+              std::make_unique<WasmEdgeTensorflowGetOutputTensor>());
+  addHostFunc("wasmedge_tensorflow_get_tensor_len",
+              std::make_unique<WasmEdgeTensorflowGetTensorLen>());
+  addHostFunc("wasmedge_tensorflow_get_tensor_data",
+              std::make_unique<WasmEdgeTensorflowGetTensorData>());
+  addHostFunc("wasmedge_tensorflow_append_input",
+              std::make_unique<WasmEdgeTensorflowAppendInput>());
+  addHostFunc("wasmedge_tensorflow_append_output",
+              std::make_unique<WasmEdgeTensorflowAppendOutput>());
+  addHostFunc("wasmedge_tensorflow_clear_input",
+              std::make_unique<WasmEdgeTensorflowClearInput>());
+  addHostFunc("wasmedge_tensorflow_clear_output",
+              std::make_unique<WasmEdgeTensorflowClearOutput>());
 }
 
 } // namespace Host
-} // namespace SSVM
+} // namespace WasmEdge