+++ lib/tensorflowlite_module.cpp
@@ -4,26 +4,26 @@
 
 #include <memory>
 
-namespace SSVM {
+namespace WasmEdge {
 namespace Host {
 
-SSVMTensorflowLiteModule::SSVMTensorflowLiteModule()
-    : ImportObject("ssvm_tensorflowlite") {
-  addHostFunc("ssvm_tensorflowlite_create_session",
-              std::make_unique<SSVMTensorflowLiteCreateSession>());
-  addHostFunc("ssvm_tensorflowlite_delete_session",
-              std::make_unique<SSVMTensorflowLiteDeleteSession>());
-  addHostFunc("ssvm_tensorflowlite_run_session",
-              std::make_unique<SSVMTensorflowLiteRunSession>());
-  addHostFunc("ssvm_tensorflowlite_get_output_tensor",
-              std::make_unique<SSVMTensorflowLiteGetOutputTensor>());
-  addHostFunc("ssvm_tensorflowlite_get_tensor_len",
-              std::make_unique<SSVMTensorflowLiteGetTensorLen>());
-  addHostFunc("ssvm_tensorflowlite_get_tensor_data",
-              std::make_unique<SSVMTensorflowLiteGetTensorData>());
-  addHostFunc("ssvm_tensorflowlite_append_input",
-              std::make_unique<SSVMTensorflowLiteAppendInput>());
+WasmEdgeTensorflowLiteModule::WasmEdgeTensorflowLiteModule()
+    : ImportObject("wasmedge_tensorflowlite") {
+  addHostFunc("wasmedge_tensorflowlite_create_session",
+              std::make_unique<WasmEdgeTensorflowLiteCreateSession>());
+  addHostFunc("wasmedge_tensorflowlite_delete_session",
+              std::make_unique<WasmEdgeTensorflowLiteDeleteSession>());
+  addHostFunc("wasmedge_tensorflowlite_run_session",
+              std::make_unique<WasmEdgeTensorflowLiteRunSession>());
+  addHostFunc("wasmedge_tensorflowlite_get_output_tensor",
+              std::make_unique<WasmEdgeTensorflowLiteGetOutputTensor>());
+  addHostFunc("wasmedge_tensorflowlite_get_tensor_len",
+              std::make_unique<WasmEdgeTensorflowLiteGetTensorLen>());
+  addHostFunc("wasmedge_tensorflowlite_get_tensor_data",
+              std::make_unique<WasmEdgeTensorflowLiteGetTensorData>());
+  addHostFunc("wasmedge_tensorflowlite_append_input",
+              std::make_unique<WasmEdgeTensorflowLiteAppendInput>());
 }
 
 } // namespace Host
-} // namespace SSVM
+} // namespace WasmEdge