+++ Changelog.md
@@ -1,3 +1,18 @@
+### 0.8.0-rc2 (2021-05-07)
+
+This is the host function extension for [WasmEdge](https://github.com/WasmEdge/WasmEdge).
+Please refer to the [WasmEdge 0.8.0](https://github.com/WasmEdge/WasmEdge/releases/tag/0.8.0) for more details.
+
+Features:
+
+* Renamed this project to `WasmEdge-tensorflow` and updated the `WasmEdge` dependency.
+* Add `wasmedge-tensorflow` and `wasmedge-tensorflowlite` C API shared library.
+* Add release CI.
+
+Tools:
+
+* Remove tools. The tools will be in the new [repository](https://github.com/second-state/WasmEdge-tensorflow-tools).
+
 ### 0.8.0-rc1 (2021-04-30)
 
 This is the host function extension for [SSVM](https://github.com/second-state/SSVM).